refactor(core): use gql.tada on remaining queries
