+++ apps/core/client/fragments/page-details.ts
@@ -1,8 +1,10 @@
-export const PAGE_DETAILS_FRAGMENT = /* GraphQL */ `
+import { graphql } from '../graphql';
+
+export const PAGE_DETAILS_FRAGMENT = graphql(`
   fragment PageDetails on PageInfo {
     hasNextPage
     hasPreviousPage
     startCursor
     endCursor
   }
-`;
+`);
