+++ apps/core/client/queries/get-featured-products.ts
@@ -4,22 +4,26 @@ import { cache } from 'react';
 import { getSessionCustomerId } from '~/auth';
 
 import { client } from '..';
-import { graphql } from '../generated';
+import { PRODUCT_DETAILS_FRAGMENT } from '../fragments/product-details';
+import { graphql } from '../graphql';
 import { revalidate } from '../revalidate-target';
 
-export const GET_FEATURED_PRODUCTS_QUERY = /* GraphQL */ `
-  query getFeaturedProducts($first: Int, $imageHeight: Int!, $imageWidth: Int!) {
-    site {
-      featuredProducts(first: $first) {
-        edges {
-          node {
-            ...ProductDetails
+const GET_FEATURED_PRODUCTS_QUERY = graphql(
+  `
+    query getFeaturedProducts($first: Int, $imageHeight: Int!, $imageWidth: Int!) {
+      site {
+        featuredProducts(first: $first) {
+          edges {
+            node {
+              ...ProductDetails
+            }
           }
         }
       }
     }
-  }
-`;
+  `,
+  [PRODUCT_DETAILS_FRAGMENT],
+);
 
 interface Options {
   first?: number;
@@ -29,11 +33,10 @@ interface Options {
 
 export const getFeaturedProducts = cache(
   async ({ first = 12, imageHeight = 300, imageWidth = 300 }: Options = {}) => {
-    const query = graphql(GET_FEATURED_PRODUCTS_QUERY);
     const customerId = await getSessionCustomerId();
 
     const response = await client.fetch({
-      document: query,
+      document: GET_FEATURED_PRODUCTS_QUERY,
       variables: { first, imageWidth, imageHeight },
       customerId,
       fetchOptions: customerId ? { cache: 'no-store' } : { next: { revalidate } },
