+++ apps/core/client/queries/get-newest-products.ts
@@ -4,22 +4,26 @@ import { cache } from 'react';
 import { getSessionCustomerId } from '~/auth';
 
 import { client } from '..';
-import { graphql } from '../generated';
+import { PRODUCT_DETAILS_FRAGMENT } from '../fragments/product-details';
+import { graphql } from '../graphql';
 import { revalidate } from '../revalidate-target';
 
-export const GET_NEWEST_PRODUCTS_QUERY = /* GraphQL */ `
-  query getNewestProducts($first: Int, $imageHeight: Int!, $imageWidth: Int!) {
-    site {
-      newestProducts(first: $first) {
-        edges {
-          node {
-            ...ProductDetails
+const GET_NEWEST_PRODUCTS_QUERY = graphql(
+  `
+    query getNewestProducts($first: Int, $imageHeight: Int!, $imageWidth: Int!) {
+      site {
+        newestProducts(first: $first) {
+          edges {
+            node {
+              ...ProductDetails
+            }
           }
         }
       }
     }
-  }
-`;
+  `,
+  [PRODUCT_DETAILS_FRAGMENT],
+);
 
 interface Options {
   first?: number;
@@ -29,11 +33,10 @@ interface Options {
 
 export const getNewestProducts = cache(
   async ({ first = 12, imageHeight = 300, imageWidth = 300 }: Options = {}) => {
-    const query = graphql(GET_NEWEST_PRODUCTS_QUERY);
     const customerId = await getSessionCustomerId();
 
     const response = await client.fetch({
-      document: query,
+      document: GET_NEWEST_PRODUCTS_QUERY,
       variables: { first, imageWidth, imageHeight },
       customerId,
       fetchOptions: customerId ? { cache: 'no-store' } : { next: { revalidate } },
