+++ apps/core/client/queries/get-related-products.ts
@@ -4,32 +4,36 @@ import { cache } from 'react';
 import { getSessionCustomerId } from '~/auth';
 
 import { client } from '..';
-import { graphql } from '../generated';
+import { PRODUCT_DETAILS_FRAGMENT } from '../fragments/product-details';
+import { graphql } from '../graphql';
 import { revalidate } from '../revalidate-target';
 
 import { GetProductOptions } from './get-product';
 
-export const GET_RELATED_PRODUCTS = /* GraphQL */ `
-  query getRelatedProducts(
-    $entityId: Int!
-    $optionValueIds: [OptionValueId!]
-    $first: Int!
-    $imageHeight: Int!
-    $imageWidth: Int!
-  ) {
-    site {
-      product(entityId: $entityId, optionValueIds: $optionValueIds) {
-        relatedProducts(first: $first) {
-          edges {
-            node {
-              ...ProductDetails
+const GET_RELATED_PRODUCTS = graphql(
+  `
+    query getRelatedProducts(
+      $entityId: Int!
+      $optionValueIds: [OptionValueId!]
+      $first: Int!
+      $imageHeight: Int!
+      $imageWidth: Int!
+    ) {
+      site {
+        product(entityId: $entityId, optionValueIds: $optionValueIds) {
+          relatedProducts(first: $first) {
+            edges {
+              node {
+                ...ProductDetails
+              }
             }
           }
         }
       }
     }
-  }
-`;
+  `,
+  [PRODUCT_DETAILS_FRAGMENT],
+);
 
 export const getRelatedProducts = cache(
   async (
@@ -41,11 +45,10 @@ export const getRelatedProducts = cache(
   ) => {
     const { productId, optionValueIds, first = 12, imageWidth = 300, imageHeight = 300 } = options;
 
-    const query = graphql(GET_RELATED_PRODUCTS);
     const customerId = await getSessionCustomerId();
 
     const response = await client.fetch({
-      document: query,
+      document: GET_RELATED_PRODUCTS,
       variables: { entityId: productId, optionValueIds, first, imageWidth, imageHeight },
       fetchOptions: customerId ? { cache: 'no-store' } : { next: { revalidate } },
     });
